amendment title in DIS: https://github.com/metanorma/metanorma-iso/issues/675
